Create StateEstimationService class and move function from studyService to it
